LayoutTests/http/tests/security/contentSecurityPolicy/eval-blocked-in-about-blank-iframe-expected.txt

ALERT: /PASS/
CONSOLE MESSAGE: line 1: EvalError: Eval is disabled
 Eval should be blocked in the iframe, but inline script should be allowed.
